requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.2.2
idna==2.6
pyasn1==0.4.2
pycparser==2.18
pyOpenSSL==17.5.0
rsa==3.4.2
six==1.11.0
